refactor(core): move RowPositionIndexer into slick.core.ts
The class was the only content of slick.rowpositionindex.ts and is
consumed solely by the grid, so it now lives in slick.core.ts alongside
the other core classes (its exports land on the Slick namespace in the
iife build automatically).

This removes the feature's entire extra deployment surface: no new
dist/browser artifact, no script-tag ordering requirement, and the
loader-guard error is gone (slick.core.js always precedes
slick.grid.js). Examples drop the extra script tag; the esm/cjs bundles
and typings are unchanged in content.
