+++ examples/example-variable-row-height-frozen.html
@@ -46,8 +46,6 @@
 
 <script src="../dist/browser/slick.core.js"></script>
 <script src="../dist/browser/slick.interactions.js"></script>
-<!-- slick.rowpositionindex.js is required when using rowHeightProvider and must be loaded before slick.grid.js -->
-<script src="../dist/browser/slick.rowpositionindex.js"></script>
 <script src="../dist/browser/slick.grid.js"></script>
 
 <script>
