+++ examples/example-variable-row-height-spans.html
@@ -40,8 +40,6 @@
 
 <script src="../dist/browser/slick.core.js"></script>
 <script src="../dist/browser/slick.interactions.js"></script>
-<!-- slick.rowpositionindex.js is required when using rowHeightProvider and must be loaded before slick.grid.js -->
-<script src="../dist/browser/slick.rowpositionindex.js"></script>
 <script src="../dist/browser/slick.grid.js"></script>
 
 <script>
