+++ src/global.d.ts
@@ -16,6 +16,7 @@ import type {
   SlickGroupTotals,
   SlickRange,
   SlickCopyRange,
+  RowPositionIndexer,
   Utils,
   SelectionUtils,
   ValueFilterMode,
@@ -51,7 +52,6 @@ import type { SlickGroupItemMetadataProvider } from './slick.groupitemmetadatapr
 import type { SlickRemoteModel } from './slick.remotemodel.js';
 import type { SlickRemoteModelYahoo } from './slick.remotemodel-yahoo.js';
 import type { Draggable, MouseWheel, Resizable } from './slick.interactions.js';
-import type { RowPositionIndexer } from './slick.rowpositionindex.js';
 import type { Aggregators } from './slick.dataview.js';
 import type { Editors } from './slick.editors.js';
 import type { Formatters } from './slick.formatters.js';
